Tweak SqrtField_objectify; fix mutability issue
- Declare `SqrtField_objectify` using `BindGlobal`
- Drop `type` argument of `SqrtField_objectify`, it is always SqrtFieldType
- Remove one level of nesting for the `elm` argument to `SqrtField_objectify`:
  so instead of `SqrtField_objectify([x])`, now write `SqrtField_objectify(x)`
- Finally, don't pass an immutable list as second argument to `Objectify`. This
  is trivial to achieve thanks to the previous point.

Note that passing immutable objects as second argument to `Objectify` was
always illegal in GAP, but due to a bug this restriction was not properly
enforced. As a result, code accidentally passing in immutable objects silently
worked. Future versions of GAP will issue a warning for this, and one distant
day perhaps even an error.

In the meantime, this patch prepares corelg for this eventuality, and even if
that day never comes, it does no harm either.